+++ xmonad-manage/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## master
-- Active modules are now based on configurations.
-- X11 display configurations could be configured via `display-config.yaml`.
+- Active modules are now based on configuration in `config/active-modules.yaml`.
+- X11 display configurations could be configured via `config/display-config.yaml`.
 
 ## 0.7.0
 - Now depends on systemd; profiles operate via service files.