+++ xmonad-manage/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## master
+## 0.7.1
 - Active modules are now based on configuration in `config/active-modules.yaml`.
 - X11 display configurations could be configured via `config/display-config.yaml`.